* Handle schnorr sigs of 65 bytes (including sighash type) * Verify pubkey lengths in psbt setters * Correctly identify change output * Refactor tests
